norminette casi OK. faltan algunas funciones de parser.c + error con espacio y backslash dentro de quotes con expand
