+++ Makefile
@@ -1,5 +1,5 @@
 CC 		= cc
-CFLAGS = -Wall -Wextra -Werror -g3 -Iincludes  #-fsanitize=address
+CFLAGS = -Wall -Wextra -Werror -g3 -Iincludes  -fsanitize=address
 LDFLAGS = -lreadline
 
 GREEN = \033[32m
@@ -8,17 +8,22 @@ RESET = \033[0m
 
 NAME	:= minishell
 
-SRCS	:= ${addprefix src/, minishell.c \
+SRCS	:= ${addprefix src/, minishell.c\
 		${addprefix environement/, split_env_vars.c fill_env_struct.c env_helpers.c copy_list.c free_env_node.c} \
 		${addprefix lexer/, lexer.c tag_in_quotes.c remove_spaces.c} \
-		${addprefix parser/, parser.c tokenize_strings.c tokenize_char.c parse_redirs.c expand_var.c expand_help.c parse_cmd.c parse_errors.c check_white.c} \
+		${addprefix parser/, parser.c} \
+		${addprefix parser/parse_toks/, parse_hd.c parse_cmd.c parse_errors.c parse_utils.c} \
 		${addprefix parser/order_toks/, order_toks.c order_help.c} \
-		${addprefix exec/, execution.c execute_builtin.c init_fds.c fill_fds.c get_fds.c\
-		ft_dup.c dup_tubes.c cmd_exec.c   cmd_exec_utils.c is_slash.c builtin_in_parent.c exec_in_child.c close_fd.c} \
+		${addprefix parser/expand/, expand_var.c expand_help.c expand_dq.c} \
+		${addprefix parser/tokenize/, token_ops.c tokenize_char.c tokenize_strings.c tokenize_redirs.c} \
+		${addprefix exec/, execution.c } \
+		${addprefix exec/child/, ft_dup.c dup_tubes.c cmd_exec.c cmd_exec_utils.c is_slash.c exec_in_child.c} \
+		${addprefix exec/fd/, init_fds.c fill_fds.c get_fds.c close_fd.c} \
+		${addprefix exec/exec_builtin/, builtin_in_parent.c execute_builtin.c} \
 		${addprefix builtins/, env.c exit.c pwd.c echo.c unset.c} \
 		${addprefix builtins/export/, export.c add_var_to_list.c list_to_tab.c} \
 		${addprefix builtins/cd/, cd.c go_back.c} \
-		${addprefix utils/, ft_error.c  clean_minishell.c ft_strjoin_free.c token_list.c handle_signals.c}}
+		${addprefix utils/, ft_error.c clean_minishell.c handle_signals.c}}
 
 OBJS     = $(SRCS:.c=.o)
