+++ src/builtins/echo.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   echo.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: derjavec <derjavec@student.42.fr>          +#+  +:+       +#+        */
+/*   By: sandra <sandra@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/15 15:52:38 by skanna            #+#    #+#             */
-/*   Updated: 2024/07/22 09:28:29 by derjavec         ###   ########.fr       */
+/*   Updated: 2024/07/26 13:13:05 by sandra           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,7 +42,7 @@ static int	echo_opt(t_token *cur, int *n)
 		{
 			if (check_after_dash(cur, n, &i) < 0)
 				break ;
-		}		
+		}
 		else
 		{
 			i++;
