+++ src/exec/fd/init_fds.c
@@ -1,16 +1,32 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   init_exec.c                                        :+:      :+:    :+:   */
+/*   init_fds.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: deniseerjavec <deniseerjavec@student.42    +#+  +:+       +#+        */
+/*   By: sandra <sandra@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/08 17:03:22 by skanna            #+#    #+#             */
-/*   Updated: 2024/07/10 17:27:52 by deniseerjav      ###   ########.fr       */
+/*   Updated: 2024/07/26 15:38:52 by sandra           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
+
+static int	ft_pipe_count(t_mini *mini)
+{
+	t_token	*tmp;
+	int		count;
+
+	tmp = mini->token;
+	count = 0;
+	while (tmp)
+	{
+		if (tmp->type == PIPE)
+			count++;
+		tmp = tmp->next;
+	}
+	return (count + 1);
+}
 
 static int	ft_cmd_count(t_mini *mini)
 {
@@ -50,17 +66,12 @@ static void	init_fds(t_mini *mini, int j)
 
 int	malloc_fds(t_mini *mini)
 {
-	int	i;
-	int	j;
-
+	mini->cmd_count = ft_cmd_count(mini);
 	mini->pipe_count = ft_pipe_count(mini);
-	j = mini->pipe_count;
-	if (mini->pipe_count == 0)
-		j = 1;
-	mini->fd_in = malloc(j * sizeof(int));
-	mini->fd_out = malloc(j * sizeof(int));
-	mini->inv_fd = malloc(j * sizeof(int));
-	mini->pid = malloc(j * sizeof(pid_t));
+	mini->fd_in = malloc(mini->pipe_count * sizeof(int));
+	mini->fd_out = malloc(mini->pipe_count * sizeof(int));
+	mini->inv_fd = malloc(mini->pipe_count * sizeof(int));
+	mini->pid = malloc(mini->pipe_count * sizeof(pid_t));
 	if (!mini->fd_in || !mini->fd_out || !mini->pid || !mini->inv_fd)
 		return (ft_error(mini, NULL, strerror(errno)), 1);
 	if (mini->pipe_count > 1)
@@ -69,6 +80,6 @@ int	malloc_fds(t_mini *mini)
 		if (!mini->tube)
 			return (ft_error(mini, NULL, strerror(errno)), 1);
 	}
-	init_fds(mini, j);
+	init_fds(mini, mini->pipe_count);
 	return (0);
 }
