+++ src/parser/parse_toks/parse_cmd.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parse_cmd.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: deniseerjavec <deniseerjavec@student.42    +#+  +:+       +#+        */
+/*   By: sandra <sandra@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/08 17:47:03 by sandra            #+#    #+#             */
-/*   Updated: 2024/07/10 21:59:45 by deniseerjav      ###   ########.fr       */
+/*   Updated: 2024/07/26 15:54:35 by sandra           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -81,16 +81,27 @@ static int	count_cmd_tokens(t_token *tmp)
 	return (i);
 }
 
+static void	update_tok(t_mini *mini, t_token **c, t_token **p, t_token **new)
+{
+	if (*p)
+		(*p)->next = *new;
+	else
+		mini->token = *new;
+	(*new)->next = *c;
+	*p = *new;
+}
+
 static void	create_cmd_tab(t_mini *mini, t_token **cur, t_token **prev)
 {
 	t_token	*new;
 	int		tok_count;
 
-	if (mini->token == *cur && (*cur)->value[0] == '.' && ft_strlen((*cur)->value) == 1 && !(*cur)->next)
+	if (mini->token == *cur && (*cur)->value[0] == '.' \
+	&& ft_strlen((*cur)->value) == 1 && !(*cur)->next)
 	{
 		mini->exit_status = 2;
 		return (ft_error(mini, ". :filename argument required", NULL));
-	}		
+	}
 	new = tok_new_node("", COMMAND);
 	if (!new)
 		return (ft_error(mini, NULL, strerror(errno)));
@@ -104,12 +115,7 @@ static void	create_cmd_tab(t_mini *mini, t_token **cur, t_token **prev)
 		ft_error(mini, NULL, strerror(errno));
 		return ;
 	}
-	if (*prev)
-		(*prev)->next = new;
-	else
-		mini->token = new;
-	new->next = *cur;
-	*prev = new;
+	update_tok(mini, cur, prev, &new);
 }
 
 int	parse_commands(t_mini *mini)
