+++ src/parser/parse_toks/parse_utils.c
@@ -1,39 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   check_white.c                                      :+:      :+:    :+:   */
+/*   check_white_and_slash.c                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: derjavec <derjavec@student.42.fr>          +#+  +:+       +#+        */
+/*   By: sandra <sandra@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/05 12:03:59 by skanna            #+#    #+#             */
-/*   Updated: 2024/07/22 10:07:04 by derjavec         ###   ########.fr       */
+/*   Updated: 2024/07/26 15:26:26 by sandra           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-void	join_tok(t_mini *mini, t_token **tmp, t_token **prev)
-{
-	char	*new_token;
-	t_token	*new;
-
-	new_token = ft_strjoin_free((*tmp)->value, (*tmp)->next->value);
-	if (!new_token)
-		return (ft_error(mini, NULL, strerror(errno)));
-	new = tok_new_node(new_token, STRING);
-	free(new_token);
-	if (!new)
-		return (ft_error(mini, NULL, strerror(errno)));
-	new->next = (*tmp)->next->next;
-	free((*tmp)->next);
-	free(*tmp);
-	*tmp = new;
-	if (*prev)
-		(*prev)->next = new;
-	else
-		mini->token = new;
-}
-
 
 static void	erase_white_tok(t_mini *mini, t_token **tmp, t_token **prev)
 {
