+++ src/parser/tokenize/token_ops.c
@@ -1,17 +1,38 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   token_list.c                                       :+:      :+:    :+:   */
+/*   token_ops.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: derjavec <derjavec@student.42.fr>          +#+  +:+       +#+        */
+/*   By: sandra <sandra@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/15 11:58:03 by derjavec          #+#    #+#             */
-/*   Updated: 2024/07/23 11:14:32 by derjavec         ###   ########.fr       */
+/*   Updated: 2024/07/26 15:26:38 by sandra           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
+void	join_tok(t_mini *mini, t_token **tmp, t_token **prev)
+{
+	char	*new_token;
+	t_token	*new;
+
+	new_token = ft_strjoin_free((*tmp)->value, (*tmp)->next->value);
+	if (!new_token)
+		return (ft_error(mini, NULL, strerror(errno)));
+	new = tok_new_node(new_token, STRING);
+	free(new_token);
+	if (!new)
+		return (ft_error(mini, NULL, strerror(errno)));
+	new->next = (*tmp)->next->next;
+	free((*tmp)->next);
+	free(*tmp);
+	*tmp = new;
+	if (*prev)
+		(*prev)->next = new;
+	else
+		mini->token = new;
+}
 
 void	tok_delone(t_token **lst, t_token **prev)
 {
