+++ src/parser/tokenize/tokenize_char.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   tokenize_char.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: derjavec <derjavec@student.42.fr>          +#+  +:+       +#+        */
+/*   By: sandra <sandra@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/05 12:03:59 by skanna            #+#    #+#             */
-/*   Updated: 2024/07/25 14:46:30 by derjavec         ###   ########.fr       */
+/*   Updated: 2024/07/26 15:22:29 by sandra           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
