+++ src/parser/tokenize/tokenize_redirs.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   parse_redirs.c                                     :+:      :+:    :+:   */
+/*   tokenize_redirs.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: derjavec <derjavec@student.42.fr>          +#+  +:+       +#+        */
+/*   By: sandra <sandra@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/06/09 12:04:54 by sandra            #+#    #+#             */
-/*   Updated: 2024/07/23 11:15:48 by derjavec         ###   ########.fr       */
+/*   Created: 2024/07/26 15:22:09 by sandra            #+#    #+#             */
+/*   Updated: 2024/07/26 15:24:08 by sandra           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -70,51 +70,4 @@ void	tokenize_redirs(t_mini *ms, t_pretok **cur, t_token **list)
 		else
 			break ;
 	}
-}
-
-static void	prep_lim(t_mini *mini, t_token	**cur, t_token	**prev)
-{
-	if ((*cur)->next && ((*cur)->next->type == STRING || \
-			(*cur)->next->type == D_Q || (*cur)->next->type == S_Q))
-	{
-		join_tok(mini, cur, prev);
-		(*cur)->type = LIM;
-	}
-	else
-	{
-		(*cur)->type = LIM;
-		while ((*cur) && (*cur)->type != PIPE)
-			(*cur) = (*cur)->next;
-	}
-}
-
-int	prep_heredoc(t_mini *mini)
-{
-	t_token	*prev;
-	t_token	*cur;
-
-	prev = NULL;
-	cur = mini->token;
-	while (cur)
-	{
-		if ((cur->type == STRING || cur->type == D_Q \
-			|| cur->type == S_Q || cur->type == LIM)
-			&& prev && prev->type == HERE)
-			prep_lim(mini, &cur, &prev);
-		else
-		{
-			if (cur->type != WHITE)
-			{
-				prev = cur;
-				cur = cur->next;
-			}
-			else if (cur->next && prev->type == HERE)
-				tok_delone(&cur, &prev);
-			else
-				cur = cur->next;
-		}
-		if (mini->error != 0)
-			return (1);
-	}
-	return (0);
 }
